tests/rigging/requirements.txt
# Gnosis Wraith API v2 Test Suite Requirements

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0  # For parallel test execution

# HTTP client
requests>=2.31.0
aiohttp>=3.8.0  # For async requests

# Environment and configuration
python-dotenv>=1.0.0

# Utilities
colorama>=0.4.6  # Cross-platform colored output

# Optional: Enhanced test reporting
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# Optional: Test coverage
pytest-cov>=4.1.0

# Optional: Mock and fixtures
pytest-mock>=3.11.0